legacy/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
cairosvg>=2.7.0
numpy>=1.24.0
elasticsearch>=8.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0
mcp>=0.9.0
slowapi>=0.1.9
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-elasticsearch>=0.42b0
opentelemetry-exporter-otlp-proto-http>=1.21.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
httpx>=0.24.0
faker>=19.0.0
freezegun>=1.2.0

